refactor(agents): split Sisyphus prompts by model
